Updated download logic for new CSV format
The script logic for handling the new CSV format has been updated. The
source type can either be 'build', 'sourcecode', or 'git'. In the case
of 'git', the 'git_commit' must also be provided in the CSV. If the
pkg_url is not an http URL, the script will treat the pkg_url as a local
path to a build file. Depending on the last 4-8 characters of the
pkg_url, the script will treat the downloaded file as either a .tar.gz,
.tar.bz2, .zip, or .rar